+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.1
+
+- Fixed ttl calculating when dealing with milliseconds.
+- Fixed problems while dealing with expired items in "clearExpired" and "setMany" of `MultiTimedCacheModel`.
+
 ## 1.1.0
 
 - Added "assumeAllPresent" option to `MultiTimesCacheModel.getMany`.